Login Page
The user can log in if they have an administrator account; otherwise, they should log in as a client. The username and password are verified on the server to allow the user to use the POST, PUT, and DELETE methods. Additionally, they can get to know the brands we work with!